updated layout of report list

+++ lib/screens/report_detail.dart
@@ -212,7 +212,8 @@ class ReportDetailState extends State<ReportDetail> {
                     SizedBox(
                       width: 5,
                     ),
-                    Text(formatter.format(int.parse(report.reportEmralsAmount)))
+                    Text(formatter
+                        .format(double.parse(report.reportEmralsAmount)))
                   ]),
                   Row(
                     children: <Widget>[
